Factor LightingEffect out of Renderer (#26)
This patch factors the code for the lighting effect out of the all-sing,
all-dance "Renderer" class and into a self-contained LightingEffect
class. Also, gather all the various shaders and resources used by the
lighting effect into //escher/effects/lighting alongside LightingEffect.

Finally, fix a bug with where we were drawing the floating action button
so that it is actually centered under the user's finger.