Factor //escher/base into //ftl (#27)
The code in //escher/base isn't really related to Escher. It's the seed
of a base library for Fuchsia. Talking with trung, I think we want to
call this library FTL. We can incubate it here for a while, but it
should eventually become a separate library that can be used in more
places.